Moved database helpers to a separate package
