+++ app/src/main/java/io/github/tstewart/todayi/helpers/db/DayRatingTableHelper.java
@@ -1,4 +1,4 @@
-package io.github.tstewart.todayi.helpers;
+package io.github.tstewart.todayi.helpers.db;
 
 import android.content.Context;
 import android.database.Cursor;
@@ -9,8 +9,9 @@ import java.util.Date;
 import androidx.annotation.NonNull;
 import io.github.tstewart.todayi.data.DBConstants;
 import io.github.tstewart.todayi.data.Database;
-import io.github.tstewart.todayi.data.UserPreferences;
 import io.github.tstewart.todayi.errors.ValidationFailedException;
+import io.github.tstewart.todayi.helpers.DateFormatter;
+import io.github.tstewart.todayi.helpers.db.DatabaseHelper;
 import io.github.tstewart.todayi.models.DayRating;
 
 /**
